f64 resample ratio fixes weird phase issue

+++ generic_daw_core/src/audio_clip.rs
@@ -5,7 +5,7 @@ use dsp::resample_cubic;
 pub struct AudioClip {
 	pub sample: SampleId,
 	pub position: OffsetBeatSpan,
-	pub stretch: f32,
+	pub stretch: f64,
 }
 
 impl AudioClip {
