+++ generic_daw_gui/src/arrangement_view/sample.rs
@@ -18,8 +18,8 @@ pub struct Sample {
 }
 
 impl Sample {
-	pub fn resample_ratio(&self, transport: &Transport) -> f32 {
-		transport.sample_rate.get() as f32 / self.sample_rate.get() as f32
+	pub fn resample_ratio(&self, transport: &Transport) -> f64 {
+		f64::from(transport.sample_rate.get()) / f64::from(self.sample_rate.get())
 	}
 
 	pub fn len(&self, transport: &Transport) -> BeatTime {
